requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn>=0.23.0
duckdb>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0,<3.0.0
slowapi>=0.1.9
